Added more Ubersmith exceptions:
- Timeout Exceptions
- Connection Errors
